requirements.txt
discord.py
python-dotenv
atproto